Update net.sf.prenzl plugin

+++ Java/PrenzlRCP/plugins/net.sf.prenzl/src/net/sf/prenzl/launch/ICountListener.java
@@ -1,5 +1,5 @@
 package net.sf.prenzl.launch;
 
 public interface ICountListener {
-	public void setCount(long count);
+	public void setCount(int count);
 }
